Add custom argument representation handling
Add the ability for types to control how they are sent when used as
arguments to redis commands.

This is useful for types which don't want to use their Stringer
representation when used as an argument to a redis command.

Examples of this is a time.Duration which the user wants to store
the number of seconds or a time.Time being stored as int instead of
a string.