Treat named primitive types as primitives, rather than as strings.
As an interim measure to a more full-featured argument serializer
interface, handle simply-wrapped types as their primitives rather than
falling back to Stringer-style conversion.

This is particularly useful in the case of enum/iota-style constants
that are typed with a specific name.  Previously, any `String()` method
on this type would result in those stringified values getting written to
redis, and then a parse failure when read, especially when read via
`ScanStruct()`, where it's difficult to intervene in the parsing.  With
this change, those values are written as their primitive (numeric) type,
and thus read as that primitive type, and can be successfully assigned
to the named type even when it's a member of a struct.