Add Pool Do and DoScript helpers
Add Do and DoScript helpers to Pool which allow individual Do commands to be easily executed while holding a connection for a minimum amount of time.

This is important for apps which mix redis commands with other work, which would otherwise need to implement individual functions or closures which perform the Get / Close methods in order to avoid holding redis connections while not actually using them.